refactor: use percent for speed in gameboy.rs
feat: add 0.25, 0.5 and 1.5 speed

feat: better interface

refactor: renamed speed_percent into speed

refactor: fmt

+++ src/communications.rs
@@ -98,7 +98,7 @@ pub enum Request {
     StopWatch(u16),
     Step(usize),
     SetInstructionListLength(u8),
-    SetSpeed(u8),
+    SetSpeed(u16),
     SetVolume(u8),
 }
